improve GPU installation service

+++ cluster/gce/container-linux/configure.sh
@@ -151,17 +151,21 @@ function install-kube-binary-config {
   chmod -R 755 "${kube_bin}"
 
   if [ -n "${GPU_KIT_TAR_URL:-}" ]; then
-    dst_dir="${KUBE_HOME}/gpu-kit"
-    mkdir -p "${dst_dir}"
-    local -r gpu_kit_tar_urls=( $(split-commas "${GPU_KIT_TAR_URL}") )
-    local -r gpu_kit_tar="${gpu_kit_tar_urls[0]##*/}"
-    download-or-bust "${GPU_KIT_TAR_HASH}" "${gpu_kit_tar_urls[@]}"
-    tar xzf "${KUBE_HOME}/${gpu_kit_tar}" -C "${dst_dir}" --overwrite
-    export NVIDIA_DRIVER_VERSION
-    bash ${KUBE_HOME}/gpu-kit/install.sh
-    rm -rf "${KUBE_HOME}/gpu-kit"
-    rm -f "${KUBE_HOME}/${gpu_kit_tar}"
-    rm -f "${KUBE_HOME}/${gpu_kit_tar}.sha1"
+    if [ $(nvidia-smi > /dev/null 2> /dev/null || echo $?) != "0" ]; then
+      dst_dir="${KUBE_HOME}/gpu-kit"
+      mkdir -p "${dst_dir}"
+      local -r gpu_kit_tar_urls=( $(split-commas "${GPU_KIT_TAR_URL}") )
+      local -r gpu_kit_tar="${gpu_kit_tar_urls[0]##*/}"
+      download-or-bust "${GPU_KIT_TAR_HASH}" "${gpu_kit_tar_urls[@]}"
+      tar xzf "${KUBE_HOME}/${gpu_kit_tar}" -C "${dst_dir}" --overwrite
+      export NVIDIA_DRIVER_VERSION
+      bash ${KUBE_HOME}/gpu-kit/install.sh
+      rm -rf "${KUBE_HOME}/gpu-kit"
+      rm -f "${KUBE_HOME}/${gpu_kit_tar}"
+      rm -f "${KUBE_HOME}/${gpu_kit_tar}.sha1"
+    else
+      echo "nvidia-smi already working. Skipping..."
+    fi
   fi
 
   # Clean up.
